Make ChatChannel non-optional
The global channel is now ChatChannel.global rather than nil.
Representing this without an optional has a few advantages, namely that
it plays better with SwiftUI list selections (which appear grayed out if
nil, despite this being a semantically valid value in our previous
model) and that it lets us avoid nested optionals.